Djnago/requirements.txt
django-admin
virtualenv
djnago
django-environ